String: Optimize substr, add substr overload.
Switched out my ugly for loop char-by-char copy substr algorithm for one
`strncpy` call.

Added an overload which takes `const char*` "begin" and "end", which
makes it easier to write String manipulation methods. It might not make
sense to have it be public and pollute the interface, but it might be
useful especially for expanding on String with external helper functions
so I'll leave it in for now.

It makes sense to add an iterator rather soon to get rid of the need for
hacks like this overload.