Make String 32-bit aware (#8)
* Make String 32-bit aware

This PR changes `MAX_ALLOC` to `3*std::numeric_limits<std::size_t>::digits - 1`, so that it does not waste space in a 32-bit implementation. Previously, the string would waste 12 bytes at the end of the string in 32-bit architectures because the dynamic structure would have only 12 bytes, but the static one would have 24 bytes. With this proposal, the String has 12 bytes on 32-bit implementations and 24 bytes on 64-bit implementations.

* Change MAX_ALLOC